feat(content): 매장 위치 기반 AI 날씨 컨텍스트 생성 (S14P31A401-179)

+++ be/src/main/java/com/matketing/be/domain/content/dto/AiWeatherRequest.java
@@ -15,7 +15,7 @@ public record AiWeatherRequest(
         @JsonProperty("diurnal_range")
         Double diurnalRange,
         @JsonProperty("discomfort_index")
-        Double discomfortIndex,
+        Integer discomfortIndex,
         @JsonProperty("heavy_rain_warning")
         String heavyRainWarning,
         @JsonProperty("typhoon_warning")
